server/todos/requirements.txt

asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
Flask==2.1.1
Jinja2==3.1.1
mysqlclient==2.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.18.1
